added scaffolding for command line tool

+++ openmdao/utils/scaffold.py
@@ -19,6 +19,7 @@ _common_bases = {
     'SurrogateModel': 'openmdao_surrogate_models',
     'CaseRecorder': 'openmdao_case_recorders',
     'BaseCaseReader': 'openmdao_case_readers',
+    '@command': 'openmdao_commands',
 }
 
 
@@ -33,14 +34,16 @@ def _scaffold_setup_parser(parser):
     """
     parser.add_argument('file', nargs='?', help='output file.')
     parser.add_argument('-c', '--class', action='store', dest='class_name',
-                        required=True, help='Name of the class.  If an output file '
+                        help='Name of the new class.  If an output file '
                         'is not provided, this name will be used to generate the output file name.')
     parser.add_argument('-b', '--base', action='store', dest='base',
-                        required=True, help='Name of the base class for the new class. Allowed '
+                        help='Name of the base class for the new class. Allowed '
                         'base classes are: {}'.format(sorted(_common_bases)))
     parser.add_argument('-p', '--package', action='store', dest='package',
                         help="Specify name of python package.  If this is specified, the directory"
                              " structure for a python package will be created.")
+    parser.add_argument('--cmd', action='store', dest='command_name',
+                        help="Create scaffolding for an OpenMDAO command line tool.")
 
 
 def _camel_case_split(cname):
@@ -68,6 +71,8 @@ def _write_template(outfile, prefix, **kwargs):
         contents = template
 
     if outfile is not None:
+        if os.path.exists(outfile):
+            raise RuntimeError("'{}' already exists.".format(outfile))
         with open(outfile, 'w') as f:
             f.write(contents)
 
@@ -85,16 +90,29 @@ def _scaffold_exec(options, user_args):
     user_args : list of str
         Command line options after '--' (if any).  Passed to user script.
     """
-    if options.file is None:
-        outfile = _camel_case_split(options.class_name)
-    else:
-        outfile = os.path.splitext(options.file)[0]
-
+    outfile = os.path.splitext(options.file)[0] if options.file else None
     base = options.base
+    if options.command_name:
+        if base is not None or options.class_name is not None:
+            raise RuntimeError("You cannot specify (class_name, base_name) and (command_name) at "
+                               "the same time.")
+        base = '@command'
+        epname = options.command_name
+        tgtname = '_' + epname + '_setup'
+        if options.file is None:
+            outfile = _camel_case_split(options.command_name)
+    elif options.class_name is None or options.base is None:
+        raise RuntimeError("One of [--class, --base] was not specified.")
+    else:
+        epname = options.class_name.lower()
+        tgtname = options.class_name
+        if options.file is None:
+            outfile = _camel_case_split(options.class_name)
+
     start_dir = os.getcwd()
 
-    compfile = outfile + '.py'
-    testfile = 'test_' + compfile
+    pyfile = outfile + '.py'
+    testfile = 'test_' + pyfile
 
     if base in _common_bases:
         if options.package:  # create a package
@@ -105,8 +123,8 @@ def _scaffold_exec(options, user_args):
             else:
                 keywords = []
 
-            if os.path.exists(outfile):
-                raise RuntimeError("'{}' already exists.".format(outfile))
+            if os.path.exists(dist_name):
+                raise RuntimeError("'{}' already exists.".format(dist_name))
 
             # create distribution directory
             os.mkdir(dist_name)
@@ -125,8 +143,7 @@ def _scaffold_exec(options, user_args):
                 }
 
                 if entry_pt_group:
-                    entry_pt_str = "{}={}.{}:{}".format(options.class_name.lower(),
-                                                        pkg_name, outfile, options.class_name)
+                    entry_pt_str = "{}={}.{}:{}".format(epname, pkg_name, outfile, tgtname)
                     setup_dict['entry_points'] = {
                         entry_pt_group: [entry_pt_str]
                     }
@@ -141,7 +158,10 @@ def _scaffold_exec(options, user_args):
                 with open('__init__.py', 'w') as f:
                     pass
 
-                _write_template(compfile, options.base, class_name=options.class_name)
+                if base == '@command':
+                    _write_template(pyfile, 'command', command_name=options.command_name)
+                else:
+                    _write_template(pyfile, base, class_name=options.class_name)
 
                 os.mkdir('test')
                 os.chdir('test')
@@ -155,7 +175,7 @@ def _scaffold_exec(options, user_args):
             finally:
                 os.chdir(start_dir)
         else:
-            _write_template(compfile, options.base, class_name=options.class_name)
+            _write_template(pyfile, base, class_name=options.class_name)
             _write_template(testfile, 'test', class_name=options.class_name)
     else:
         raise RuntimeError("Unrecognized base class '{}'.".format(base))
